feat: #WB-2391, refactor as a simple app and an Editor component + context + hook
